Full Rewrite
Fully rewrote the whole plugin to make it more flexible, efficient, modular, and overall just better.
Now, adding a turn is as simple as just copy-pasting a class and changing the implementations.
Adding a whole new Ruleset is also as simple as just copy-pasting a class and changing some overrides and the GameInfo object.