Database Defaults Fix
> Fixes #14

The default value of the cosmetic place sound is now Sound#BLOCK_NOTE_BLOCK_BELL, instead of PlaceParticle#ENCHANTMENT, which caused errors, as it was expecting a sound name, but got a place particle. This only happened on a failed database connection or when the database was disabled.

Also made the turn countdown be checked for equality instead of equal-or-less, to allow for future games to use a turn countdown of -1, completely disabling it.