Update Dependencies & Plugins
Added paperweight-userdev for interacting with Minecraft internals for future usage. This also tightens the compatibility with other Minecraft versions.
This has also required the addition of a NeoLoader class.

Updated some parts in the build.gradle.kts to be more modern and robust.

Updated these plugins and dependencies:
- ShadowJar: 8.3.0 -> 8.3.6
- Run Paper: 2.2.3 -> 2.3.1
- Paper/Minecraft: 1.21.1 -> 1.21.4
- PostgreSQL: 42.7.4 -> 42.7.5
- InvUI: 1.38 -> 1.43
- Userdev: ? -> 2.0.0-beta.14

The new LibPG version has also allowed the removal of several utility classes like ItemBuilder, which are now part of LibPG.

Also migrated a lot of code to new Paper API features, like paintings being asset-id-based instead of a simple enum now.