Configuration Class & Colors JavaDoc
Most configuration values are now stored in a static-only Configuration class so they are only loaded once.

Also changed some values in the configuration and language file.

To top it off, I added JavaDocs to the Colors class to make some stuff more descriptive as the naming choices aren't the best.

Another small but still major change is that Commands#startCommand was renamed to Commands#gameCommand to match the literal name.