0.1.0 - Kotlin Rework
Completely moved the whole plugin from Java to Kotlin, including like 10k+ lines of code and hundreds of files.

This took a long while to do, but I believe it will be worth it making the development experience a lot better.

Please keep in mind that this is highly experimental and is not yet production-ready.