Static Structure System
The structure system now has static enums for easier access.

Not only does this come with structures in a nested enum, but also with all structures being renamed, reordered, and now actually attached to the plugin files.

Also, the arena clearing is now done for all warps that were entered, instead of only the latest one.

Other things:
- Now using Configuration.SPAWN_CORD for some stuff.
- Added a sendMessage(...) method to the Target interface, which takes the same parameters as Locale.component() and automatically translates using the target's locale(s).
- Some other smaller stuff.
- If the turning player is for some reason no longer in the game, the game will now move to the next player.