Remove Legacy Mode
Removed the legacy mode, as there's nobody who actually wants to play it, and it was buggy anyways.

Also added a new variable to the GameDifficulty for if the positions of blocks matter, like counters needing to be on top of the attacks.

And fixed the grammar inside the README and replaced the examples with legacy turns.