Hide unaffordable Turns in Shop
Added the ability to hide turns in the shop if they are too expensive for you to buy.

This can be toggled via a new button at the top left corner, and will result in the shop being-resynced a lot more.

Also made dirt and the flower pot more affordable (2 -> 1 coins).